Split front-end views and API views into separate files
